# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BookMyEvent API",
    "description": "Ticketing system API for event organisers to create & manage events.\nUsers can purchase/reserve tickets for events.\n",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@bookmyevent.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/venues": {
      "post": {
        "operationId": "VenuesController_create",
        "summary": "Create a new venue",
        "tags": ["Venues"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVenueDto" },
              "example": {
                "name": "Convention Center",
                "location": "123 Main Street, City",
                "capacity": 500
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Venue created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "Convention Center",
                    "location": "123 Main Street, City",
                    "capacity": 500,
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  },
                  "message": "Venue created successfully"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "VenuesController_findAll",
        "summary": "Get all venues",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "createdAt" }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of venues",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "Convention Center",
                      "location": "123 Main Street",
                      "capacity": 500,
                      "createdAt": "2026-04-18T12:00:00.000Z"
                    },
                    {
                      "id": "550e8400-e29b-41d4-a716-446655440001",
                      "name": "City Hall",
                      "location": "456 Downtown",
                      "capacity": 200,
                      "createdAt": "2026-04-18T11:00:00.000Z"
                    }
                  ],
                  "pagination": { "skip": 0, "take": 10, "total": 2 }
                }
              }
            }
          }
        }
      }
    },
    "/venues/{id}": {
      "get": {
        "operationId": "VenuesController_findOne",
        "summary": "Get a venue by ID",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Venue found",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "Convention Center",
                    "location": "123 Main Street",
                    "capacity": 500,
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": { "description": "Venue not found" }
        }
      },
      "patch": {
        "operationId": "VenuesController_update",
        "summary": "Update a venue",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateVenueDto" },
              "example": {
                "name": "Updated Convention Center",
                "capacity": 600
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Venue updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "Updated Convention Center",
                    "location": "123 Main Street",
                    "capacity": 600,
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  },
                  "message": "Venue updated successfully"
                }
              }
            }
          },
          "404": { "description": "Venue not found" }
        }
      },
      "delete": {
        "operationId": "VenuesController_remove",
        "summary": "Delete a venue",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "204": { "description": "Venue deleted successfully" },
          "404": { "description": "Venue not found" }
        }
      }
    },
    "/venues/{id}/seats": {
      "get": {
        "operationId": "VenuesController_getSeats",
        "summary": "Get seats for a venue",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "createdAt" }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of seats",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "aa0e8400-e29b-41d4-a716-446655440000",
                      "venueId": "550e8400-e29b-41d4-a716-446655440000",
                      "row": "A",
                      "number": 1
                    },
                    {
                      "id": "aa0e8400-e29b-41d4-a716-446655440001",
                      "venueId": "550e8400-e29b-41d4-a716-446655440000",
                      "row": "A",
                      "number": 2
                    }
                  ],
                  "pagination": { "skip": 0, "take": 50, "total": 150 }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "VenuesController_createSeat",
        "summary": "Create a seat for a venue",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VenueCreateSeatDto" },
              "example": {
                "row": "A",
                "number": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seat created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "aa0e8400-e29b-41d4-a716-446655440000",
                    "venueId": "550e8400-e29b-41d4-a716-446655440000",
                    "row": "A",
                    "number": 1
                  },
                  "message": "Seat created successfully"
                }
              }
            }
          }
        }
      }
    },
    "/venues/{id}/seats/bulk": {
      "post": {
        "operationId": "VenuesController_createBulkSeats",
        "summary": "Create bulk seats for a venue",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenueCreateBulkSeatDto"
              },
              "example": {
                "rows": ["A", "B", "C"],
                "startNumber": 1,
                "endNumber": 50
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seats created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "aa0e8400-e29b-41d4-a716-446655440000",
                      "venueId": "550e8400-e29b-41d4-a716-446655440000",
                      "row": "A",
                      "number": 1
                    },
                    {
                      "id": "aa0e8400-e29b-41d4-a716-446655440001",
                      "venueId": "550e8400-e29b-41d4-a716-446655440000",
                      "row": "A",
                      "number": 2
                    }
                  ],
                  "message": "150 seats created successfully"
                }
              }
            }
          }
        }
      }
    },
    "/venues/{id}/events": {
      "get": {
        "operationId": "VenuesController_getEvents",
        "summary": "Get events for a venue",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "createdAt" }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "660e8400-e29b-41d4-a716-446655440000",
                      "name": "Summer Concert 2026",
                      "venueId": "550e8400-e29b-41d4-a716-446655440000",
                      "status": "DRAFT"
                    }
                  ],
                  "pagination": { "skip": 0, "take": 50, "total": 1 }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateVenueDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Venue name" },
          "location": { "type": "string", "description": "Venue location" },
          "capacity": { "type": "integer", "description": "Venue capacity" }
        },
        "required": ["name", "location", "capacity"]
      },
      "UpdateVenueDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "location": { "type": "string" },
          "capacity": { "type": "integer" }
        }
      },
      "VenueCreateSeatDto": {
        "type": "object",
        "properties": {
          "row": {
            "type": "string",
            "description": "Seat row (e.g., 'A', 'B')"
          },
          "number": { "type": "integer", "description": "Seat number" }
        },
        "required": ["row", "number"]
      },
      "VenueCreateBulkSeatDto": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Array of row identifiers (e.g., ['A', 'B', 'C'])"
          },
          "startNumber": {
            "type": "integer",
            "description": "Start seat number"
          },
          "endNumber": { "type": "integer", "description": "End seat number" }
        },
        "required": ["rows", "startNumber", "endNumber"]
      },
...
    }
  }
}